requirements.txt
streamlit==1.36.0
pandas>=2.0
numpy>=1.24
Pillow>=10.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
pypdfium2>=4.30.0
matplotlib>=3.7
seaborn>=0.13
scikit-learn>=1.4
joblib>=1.3
python-dateutil>=2.8
openpyxl>=3.1
altair>=5.0
